updated tests and addressed PR comments
