+++ helpers/mocks/vaults.mock.ts
@@ -3,6 +3,10 @@ import { MakerVaultType } from 'blockchain/calls/vaultResolver'
 import { IlkData } from 'blockchain/ilks'
 import { OraclePriceData } from 'blockchain/prices'
 import { createVault$, Vault } from 'blockchain/vaults'
+import {
+  generateRandomBigNumber,
+  generateRandomString,
+} from 'features/automation/optimization/autoTakeProfit/tests/utils'
 import { PriceInfo } from 'features/shared/priceInfo'
 import { getStateUnpacker } from 'helpers/testHelpers'
 import { one, zero } from 'helpers/zero'
@@ -32,6 +36,15 @@ interface MockInstiVaultProps {
   _charterUline$?: Observable<BigNumber>
   minActiveColRatio?: BigNumber
   originationFee?: BigNumber
+}
+
+interface MockVaultOptions {
+  atRiskLevelWarning?: boolean
+  atRiskLevelDanger?: boolean
+  underCollateralized?: boolean
+  atRiskLevelWarningAtNextPrice?: boolean
+  atRiskLevelDangerAtNextPrice?: boolean
+  underCollateralizedAtNextPrice?: boolean
 }
 
 export const DEFAULT_PROXY_ADDRESS = '0xProxyAddress'
@@ -171,4 +184,46 @@ export function mockVault$({
 
 export function mockVaults(props: MockVaultProps = {}): () => Vault {
   return getStateUnpacker(mockVault$(props).vault$)
+}
+
+export function createMockVault(options: MockVaultOptions): Vault {
+  return {
+    makerType: 'STANDARD',
+    id: generateRandomBigNumber(),
+    owner: generateRandomString(),
+    controller: '',
+    token: '',
+    ilk: generateRandomString(),
+    address: generateRandomString(),
+    lockedCollateral: generateRandomBigNumber(),
+    unlockedCollateral: generateRandomBigNumber(),
+    lockedCollateralUSD: generateRandomBigNumber(),
+    lockedCollateralUSDAtNextPrice: generateRandomBigNumber(),
+    backingCollateral: generateRandomBigNumber(),
+    backingCollateralAtNextPrice: generateRandomBigNumber(),
+    backingCollateralUSD: generateRandomBigNumber(),
+    backingCollateralUSDAtNextPrice: generateRandomBigNumber(),
+    freeCollateral: generateRandomBigNumber(),
+    freeCollateralAtNextPrice: generateRandomBigNumber(),
+    freeCollateralUSD: generateRandomBigNumber(),
+    freeCollateralUSDAtNextPrice: generateRandomBigNumber(),
+    debt: generateRandomBigNumber(),
+    debtOffset: generateRandomBigNumber(),
+    normalizedDebt: generateRandomBigNumber(),
+    availableDebt: generateRandomBigNumber(),
+    availableDebtAtNextPrice: generateRandomBigNumber(),
+    collateralizationRatio: generateRandomBigNumber(),
+    collateralizationRatioAtNextPrice: generateRandomBigNumber(),
+    liquidationPrice: generateRandomBigNumber(),
+    daiYieldFromLockedCollateral: generateRandomBigNumber(),
+
+    atRiskLevelWarning: options.atRiskLevelWarning || false,
+    atRiskLevelDanger: options.atRiskLevelDanger || false,
+    underCollateralized: options.underCollateralized || false,
+
+    atRiskLevelWarningAtNextPrice: options.atRiskLevelWarningAtNextPrice || false,
+    atRiskLevelDangerAtNextPrice: options.atRiskLevelDangerAtNextPrice || false,
+    underCollateralizedAtNextPrice: options.underCollateralizedAtNextPrice || false,
+    chainId: 5,
+  } as Vault
 }
